requirements.txt
streamlit
google-generativeai
python-dotenv
passlib
sqlalchemy-libsql
datetime
